+++ sentry/src/main/java/io/sentry/SentryEnvelopeItem.java
@@ -79,7 +79,11 @@ public final class SentryEnvelopeItem {
 
     SentryEnvelopeItemHeader itemHeader =
         new SentryEnvelopeItemHeader(
-            SentryItemType.Session, () -> cachedItem.getBytes().length, "application/json", null);
+            SentryItemType.Session,
+            () -> cachedItem.getBytes().length,
+            "application/json",
+            null,
+            null);
 
     // Don't use method reference. This can cause issues on Android
     return new SentryEnvelopeItem(itemHeader, () -> cachedItem.getBytes());
@@ -116,6 +120,7 @@ public final class SentryEnvelopeItem {
             SentryItemType.resolve(event),
             () -> cachedItem.getBytes().length,
             "application/json",
+            null,
             null);
 
     // Don't use method reference. This can cause issues on Android
@@ -153,6 +158,7 @@ public final class SentryEnvelopeItem {
             SentryItemType.UserFeedback,
             () -> cachedItem.getBytes().length,
             "application/json",
+            null,
             null);
 
     // Don't use method reference. This can cause issues on Android
@@ -236,7 +242,8 @@ public final class SentryEnvelopeItem {
             SentryItemType.Attachment,
             () -> cachedItem.getBytes().length,
             attachment.getContentType(),
-            attachment.getFilename());
+            attachment.getFilename(),
+            attachment.getAttachmentType());
 
     // Don't use method reference. This can cause issues on Android
     return new SentryEnvelopeItem(itemHeader, () -> cachedItem.getBytes());