+++ sentry/src/main/java/io/sentry/SentryEnvelopeItemHeaderAdapter.java
@@ -35,6 +35,11 @@ public final class SentryEnvelopeItemHeaderAdapter extends TypeAdapter<SentryEnv
       writer.value(value.getType().getItemType().toLowerCase(Locale.ROOT));
     }
 
+    if (value.getAttachmentType() != null) {
+      writer.name("attachment_type");
+      writer.value(value.getAttachmentType());
+    }
+
     writer.name("length");
     writer.value(value.getLength());
 
@@ -52,6 +57,7 @@ public final class SentryEnvelopeItemHeaderAdapter extends TypeAdapter<SentryEnv
     String fileName = null;
     SentryItemType type = SentryItemType.Unknown;
     int length = 0;
+    String attachmentType = null;
 
     reader.beginObject();
     while (reader.hasNext()) {
@@ -75,6 +81,9 @@ public final class SentryEnvelopeItemHeaderAdapter extends TypeAdapter<SentryEnv
         case "length":
           length = reader.nextInt();
           break;
+        case "attachment_type":
+          attachmentType = reader.nextString();
+          break;
         default:
           reader.skipValue();
           break;
@@ -82,6 +91,6 @@ public final class SentryEnvelopeItemHeaderAdapter extends TypeAdapter<SentryEnv
     }
     reader.endObject();
 
-    return new SentryEnvelopeItemHeader(type, length, contentType, fileName);
+    return new SentryEnvelopeItemHeader(type, length, contentType, fileName, attachmentType);
   }
 }